refactor(envoy-filter): move filter to pkg/envoy_filter
